requirements.txt
requests
beautifulsoup4
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
selenium
webdriver-manager
python-dotenv
